Delete SidebarSelectionInfo
